# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trustleader.org API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReadyStatusStatus =
  (typeof ReadyStatusStatus)[keyof typeof ReadyStatusStatus];

export const ReadyStatusStatus = {
  ready: "ready",
} as const;

export interface ReadyStatus {
  status: ReadyStatusStatus;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type TrafficLight = (typeof TrafficLight)[keyof typeof TrafficLight];

export const TrafficLight = {
  red: "red",
  orange: "orange",
  green: "green",
} as const;

export interface InsuranceCompanySummary {
  id: string;
  name: string;
  slug: string;
  logo_url?: string | null;
  description?: string | null;
  terms_url?: string | null;
}

/** Aggregates from partner businesses and approved reviews (no claim-level data). */
export interface InsuranceCompanyStats {
  partner_businesses: number;
  total_reviews: number;
  average_rating?: number | null;
}

export interface InsuranceCompanyDetailResponse {
  insurance: InsuranceCompanySummary;
  stats: InsuranceCompanyStats;
}

export type ListingSource = (typeof ListingSource)[keyof typeof ListingSource];

export const ListingSource = {
  owner_claimed: "owner_claimed",
  import_scrape: "import_scrape",
  consumer_first_review: "consumer_first_review",
} as const;

export type ListingStatus = (typeof ListingStatus)[keyof typeof ListingStatus];

export const ListingStatus = {
  pending_verification: "pending_verification",
  active: "active",
  archived: "archived",
} as const;

export interface Business {
  id: string;
  domain: string;
  name: string;
  description?: string | null;
  traffic_light: TrafficLight;
  /** Insurer has acknowledged coverage via TrustLeader */
  insurance_proof: boolean;
  listing_source: ListingSource;
  listing_status: ListingStatus;
  insurance?: InsuranceCompanySummary | null;
  review_count: number;
  average_rating?: number | null;
  created_at: string;
  updated_at: string;
}

export interface BusinessListResponse {
  businesses: Business[];
  total: number;
  page: number;
  limit: number;
}

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus];

export const ReviewStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Review {
  id: string;
  business_id: string;
  user_id: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  text: string;
  status: ReviewStatus;
  company_response?: string | null;
  /** Present on some list endpoints (e.g. admin, consumer) */
  business_name?: string | null;
  created_at: string;
  updated_at: string;
}

/**
 * At least one field should be set
 */
export interface PatchCompanyBusinessRequest {
  name?: string;
  description?: string | null;
}

export interface AdminSetReviewStatusRequest {
  status: ReviewStatus;
}

export interface ReviewListResponse {
  reviews: Review[];
  total: number;
  page: number;
  limit: number;
}

export interface RecentReviewPublic {
  id: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  text: string;
  reviewer_label: string;
  business_id: string;
  business_name: string;
  business_domain: string;
  created_at: string;
}

export interface RecentReviewsResponse {
  reviews: RecentReviewPublic[];
  /** When the in-memory cache was last refreshed from the database */
  refreshed_at?: string;
}

export interface CreateReviewRequest {
  business_id: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @minLength 10 */
  text: string;
}

export interface UpdateReviewRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @minLength 10 */
  text: string;
}

export interface RespondToReviewRequest {
  review_id: string;
  /** @minLength 1 */
  response: string;
}

export interface ClaimBusinessRequest {
  /** Existing business UUID to claim as owner */
  business_id: string;
}

export interface ResellerStats {
  total_referrals: number;
  approved_referrals: number;
  pending_earnings: number;
  total_earnings: number;
}

export type ReferralStatus =
  (typeof ReferralStatus)[keyof typeof ReferralStatus];

export const ReferralStatus = {
  pending: "pending",
  approved: "approved",
  paid: "paid",
} as const;

export interface Referral {
  id: string;
  business_id: string;
  business_name: string;
  business_domain: string;
  status: ReferralStatus;
  commission_amount: number;
  created_at: string;
}

export interface ReferralListResponse {
  referrals: Referral[];
  total: number;
  page: number;
  limit: number;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  consumer: "consumer",
  company: "company",
  reseller: "reseller",
  admin: "admin",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  role: UserRole;
  business_id?: string | null;
  reseller_id?: string | null;
  created_at: string;
}

export type ListBusinessesParams = {
  /**
   * Search query (name or domain)
   */
  q?: string;
  page?: number;
  limit?: number;
};

export type ListInsuranceCompanyBusinessesParams = {
  page?: number;
  limit?: number;
};

export type GetBusinessReviewsParams = {
  page?: number;
  limit?: number;
};

export type GetCompanyReviewsParams = {
  status?: GetCompanyReviewsStatus;
  page?: number;
  limit?: number;
};

export type GetCompanyReviewsStatus =
  (typeof GetCompanyReviewsStatus)[keyof typeof GetCompanyReviewsStatus];

export const GetCompanyReviewsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type AdminListBusinessesParams = {
  q?: string;
  page?: number;
  limit?: number;
};

export type AdminListReviewsParams = {
  status?: ReviewStatus;
  business_id?: string;
  page?: number;
  limit?: number;
};

export type GetConsumerReviewsParams = {
  page?: number;
  limit?: number;
};

export type GetResellerReferralsParams = {
  page?: number;
  limit?: number;
};