# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trustleader.org API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /insurance-companies/{slug}:
    get:
      operationId: getInsuranceCompany
      tags: [insurance-companies]
      summary: Get insurance company by slug with aggregate stats
      description: |
        Stats are derived from active partner businesses (insurance_proof, active listing) and approved reviews.
        Insurance claim outcomes are not stored; successful-claims counts are not included.
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Insurer profile and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceCompanyDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /insurance-companies/{slug}/businesses:
    get:
      operationId: listInsuranceCompanyBusinesses
      tags: [insurance-companies]
      summary: List partner businesses for an insurance company
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated businesses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessListResponse"
        "404":
          description: Insurance company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    TrafficLight:
      type: string
      enum: [red, orange, green]
    InsuranceCompanySummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logo_url:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        terms_url:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
    InsuranceCompanyStats:
      type: object
      description: Aggregates from partner businesses and approved reviews (no claim-level data).
      properties:
        partner_businesses:
          type: integer
          description: Businesses with this insurer, insurance_proof true, and active listing
        total_reviews:
          type: integer
          description: Count of approved reviews across those businesses
        average_rating:
          type: number
          format: double
          nullable: true
          description: Mean rating over approved reviews (null if none)
      required:
        - partner_businesses
        - total_reviews
    InsuranceCompanyDetailResponse:
      type: object
      properties:
        insurance:
          $ref: "#/components/schemas/InsuranceCompanySummary"
        stats:
          $ref: "#/components/schemas/InsuranceCompanyStats"
      required:
        - insurance
        - stats
    ListingSource:
      type: string
      enum: [owner_claimed, import_scrape, consumer_first_review]
    ListingStatus:
      type: string
      enum: [pending_verification, active, archived]
    Business:
      type: object
      properties:
        id:
          type: string
        domain:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        traffic_light:
          $ref: "#/components/schemas/TrafficLight"
        insurance_proof:
          type: boolean
          description: Insurer has acknowledged coverage via TrustLeader
        listing_source:
          $ref: "#/components/schemas/ListingSource"
        listing_status:
          $ref: "#/components/schemas/ListingStatus"
        insurance:
          nullable: true
          $ref: "#/components/schemas/InsuranceCompanySummary"
        review_count:
          type: integer
        average_rating:
          type: number
          format: float
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - domain
        - name
        - traffic_light
        - insurance_proof
        - listing_source
        - listing_status
        - review_count
        - created_at
        - updated_at
    BusinessListResponse:
      type: object
      properties:
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/Business"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - businesses
        - total
        - page
        - limit
...